requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.15,<3.16
flake8>=7.0.0,<8.0.0
psycopg2-binary>=2.9,<2.10
drf-spectacular>=0.27,<0.28
Pillow>=8.2.0,<8.3.0

